Optimizing average method
Optimizing rate method by sending back the average.

+++ Functions/RateEvening.ts
@@ -77,8 +77,26 @@ function rateEvening(req: any, res: any, localPath: string) {
 
 	Rates.push(new Rate(r, ip, req.body.pc));
 
+	// Get average rate
+	if (!fs.existsSync(localPath + `rates/${month}/${day}.json`)) {
+		res.status(404).json({ error: 1, msg: 'Rates not found' });
+		return;
+	}
+	const rates = JSON.parse(fs.readFileSync(localPath + `ratesEvening/${month}/${day}.json`).toString());
+
+	let average = 0;
+
+	rates.forEach((rate: { rate: number }) => {
+		average += rate.rate;
+	});
+
+	average += r;
+
+	average = parseFloat((average / (rates.length + 1)).toFixed(1));
+
+	res.status(200).json({ error: 0, msg: 'Success', rate: average });
+
 	fs.writeFileSync(localPath + `ratesEvening/${month}/${day}.json`, JSON.stringify(Rates));
-	res.status(200).json({ error: 0, msg: 'Success' });
 }
 
 export default rateEvening;
